@@ -3,15 +3,15 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^0.7.0
-  cloud_firestore: ^0.16.0+1
   firefly:
     path: ../
+  cloud_firestore: ^2.2.0
+  firebase_core: ^1.2.0
 
 dev_dependencies:
   flutter_test:
